pyproject.toml
[project]
name = "gen-ai-practise"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "google-genai>=1.67.0",
    "ollama>=0.6.1",
    "openai>=2.28.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "tiktoken>=0.12.0",
]
